测试接口

+++ internal/user/service/user_service.go
@@ -14,8 +14,8 @@ import (
 type UserService interface {
 	SignUp(ctx context.Context, user *model.User) error
 	Login(ctx context.Context, user *model.User) (*model.User, error)
-	GetProfile(ctx context.Context, uid int64) (*model.User, error)
-	GetPermCode(ctx context.Context, uid int64) ([]string, error)
+	GetProfile(ctx context.Context, uid uint) (*model.User, error)
+	GetPermCode(ctx context.Context, uid uint) ([]string, error)
 }
 
 type userService struct {
@@ -55,11 +55,11 @@ func (us *userService) Login(ctx context.Context, user *model.User) (*model.User
 	return u, nil
 }
 
-func (us *userService) GetProfile(ctx context.Context, uid int64) (*model.User, error) {
+func (us *userService) GetProfile(ctx context.Context, uid uint) (*model.User, error) {
 	return us.dao.GetUserByID(ctx, uid)
 }
 
-func (us *userService) GetPermCode(ctx context.Context, uid int64) ([]string, error) {
+func (us *userService) GetPermCode(ctx context.Context, uid uint) ([]string, error) {
 	codes, err := us.dao.GetPermCode(ctx, uid)
 	if err != nil {
 		return nil, err
